DenCaoos-Creador

+++ commands/stickers/sticker.js
@@ -2,44 +2,41 @@ const fs = require("fs");
 
 module.exports = {
   command: ["sticker", "s"],
-  description: "Crea una imagen a sticker",
+  description: "Convierte tu basura de imagen en un sticker",
   category: "stickers",
   run: async (client, m) => {
     const quoted = m.quoted || m;
     const mime = (quoted.msg || quoted).mimetype || "";
-    const d = new Date(new Date() + 3600000);
-    const locale = "es-ES";
-    const dias = d.toLocaleDateString(locale, { weekday: "long" });
-    const fecha = d.toLocaleDateString(locale, {
-      day: "2-digit",
-      month: "2-digit",
-      year: "numeric",
-    });
-
-    let stickerTxT = "Lurus";
-    let stickerTxT2 = "Stickers";
+    
+    // Metadatos con tu Instagram
+    let pack = "CAOOS-MD PACK";
+    let autor = "IG: @DenCaoos"; // <--- Tu IG aquí
 
     if (/image/.test(mime)) {
-      media = await quoted.download();
+      m.reply("¡Cállate y espera! Estoy robando tu imagen...");
+      let media = await quoted.download();
       let encmedia = await client.sendImageAsSticker(m.chat, media, m, {
-        packname: stickerTxT2,
-        author: stickerTxT,
+        packname: pack,
+        author: autor,
       });
-      await fs.unlinkSync(encmedia);
+      if (fs.existsSync(encmedia)) await fs.unlinkSync(encmedia);
+      
     } else if (/video/.test(mime)) {
-      if ((quoted.msg || quoted).seconds > 20) {
-        return m.reply("El video no puede ser muy largo");
+      if ((quoted.msg || quoted).seconds > 11) {
+        return m.reply("¡Eres idiota! Ese video es demasiado largo.");
       }
-      media = await quoted.download();
-
+      m.reply("Procesando tu video de m***... No me presiones.");
+      let media = await quoted.download();
       let encmedia = await client.sendVideoAsSticker(m.chat, media, m, {
-        packname: "",
-        author: stickerTxT,
+        packname: pack,
+        author: autor,
       });
       await new Promise((resolve) => setTimeout(resolve, 2000));
-      await fs.unlinkSync(encmedia);
+      if (fs.existsSync(encmedia)) await fs.unlinkSync(encmedia);
+      
     } else {
-      m.reply("Envía una *imagen* o *video* junto con el comando !s");
+      m.reply("¡Aprende a usar el bot, animal! Responde a una imagen o video.");
     }
   },
 };
+
